services/document_processor/requirements.txt
# services/document_processor/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
supabase-py[async]>=2.0.0 # For updating status / potentially storing chunks metadata
pypdf2>=3.0.0
python-docx>=1.0.0  # For .docx files
python-pptx>=0.6.0 # For .pptx files



